requirements.txt
# ==============================================
# ŞIKAYET YÖNETİM SİSTEMİ - DEPENDENCIES
# ==============================================

# Web Framework
streamlit==1.31.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2  # Excel okuma/yazma için

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Type Hints Support
typing-extensions>=4.7.0

# LLM Monitoring & Evaluation
evidently>=0.4.0

# Optional: LLM Providers (Seçtiğiniz provider'a göre ekleyin)
# openai>=1.0.0
# anthropic>=0.7.0
# groq>=0.4.0
